# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tahr Air",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/flights/{flight_id}": {
      "get": {
        "tags": ["flights"],
        "summary": "Get Flight",
        "operationId": "flights-get_flight",
        "parameters": [
          {
            "name": "flight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings": {
      "post": {
        "tags": ["bookings"],
        "summary": "Create Booking",
        "operationId": "bookings-create_booking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/stats": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Stats",
        "operationId": "admin-get_stats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/flights": {
      "get": {
        "tags": ["admin"],
        "summary": "List Flights",
        "operationId": "admin-list_flights",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Flight",
        "operationId": "admin-create_flight",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/flights/{flight_id}": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Admin Flight",
        "operationId": "admin-get_admin_flight",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["admin"],
        "summary": "Update Flight",
        "operationId": "admin-update_flight",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flight Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Flight",
        "operationId": "admin-delete_flight",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/flights/{flight_id}/cancel": {
      "post": {
        "tags": ["admin"],
        "summary": "Cancel Admin Flight",
        "operationId": "admin-cancel_admin_flight",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "flight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Flight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/bookings": {
      "get": {
        "tags": ["admin"],
        "summary": "List Bookings",
        "operationId": "admin-list_bookings",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BookingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/bookings/{booking_id}": {
      "patch": {
        "tags": ["admin"],
        "summary": "Update Booking Status",
        "operationId": "admin-update_booking_status",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminStats": {
        "properties": {
          "scheduled_future_flights": {
            "type": "integer",
            "title": "Scheduled Future Flights"
          },
          "confirmed_bookings": {
            "type": "integer",
            "title": "Confirmed Bookings"
          },
          "seats_sold": {
            "type": "integer",
            "title": "Seats Sold"
          },
          "seats_available": {
            "type": "integer",
            "title": "Seats Available"
          },
          "load_factor": {
            "type": "number",
            "title": "Load Factor"
          }
        },
        "type": "object",
        "required": [
          "scheduled_future_flights",
          "confirmed_bookings",
          "seats_sold",
          "seats_available",
          "load_factor"
        ],
        "title": "AdminStats"
      },
      "AirportPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": ["id", "code", "city", "country", "name", "timezone"],
        "title": "AirportPublic"
      },
...
      "BookingCreate": {
        "properties": {
          "flight_id": {
            "type": "string",
            "format": "uuid",
            "title": "Flight Id"
          },
          "passenger_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Passenger Name"
          },
          "passenger_email": {
            "type": "string",
            "maxLength": 255,
            "format": "email",
            "title": "Passenger Email"
          }
        },
        "type": "object",
        "required": ["flight_id", "passenger_name", "passenger_email"],
        "title": "BookingCreate"
      },
...
      "BookingPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "booking_reference": {
            "type": "string",
            "title": "Booking Reference"
          },
          "flight": {
            "$ref": "#/components/schemas/FlightPublic"
          },
          "passenger_name": {
            "type": "string",
            "title": "Passenger Name"
          },
          "passenger_email": {
            "type": "string",
            "format": "email",
            "title": "Passenger Email"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "booked_price_cents": {
            "type": "integer",
            "title": "Booked Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "booking_reference",
          "flight",
          "passenger_name",
          "passenger_email",
          "status",
          "booked_price_cents",
          "currency",
          "created_at",
          "updated_at"
        ],
        "title": "BookingPublic"
      },
      "BookingStatus": {
        "type": "string",
        "enum": ["confirmed", "cancelled"],
        "title": "BookingStatus"
      },
      "BookingUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          }
        },
        "type": "object",
        "required": ["status"],
        "title": "BookingUpdate"
      },
      "BookingsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BookingPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "BookingsPublic"
      },
      "EffectiveFlightStatus": {
        "type": "string",
        "enum": ["scheduled", "cancelled", "departed"],
        "title": "EffectiveFlightStatus"
      },
      "FlightCreate": {
        "properties": {
          "flight_number": {
            "type": "string",
            "maxLength": 12,
            "minLength": 3,
            "title": "Flight Number"
          },
          "origin_id": {
            "type": "string",
            "format": "uuid",
            "title": "Origin Id"
          },
          "destination_id": {
            "type": "string",
            "format": "uuid",
            "title": "Destination Id"
          },
          "departure_at": {
            "type": "string",
            "format": "date-time",
            "title": "Departure At"
          },
          "arrival_at": {
            "type": "string",
            "format": "date-time",
            "title": "Arrival At"
          },
          "price_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Price Cents"
          },
          "total_capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Total Capacity"
          }
        },
        "type": "object",
        "required": [
          "flight_number",
          "origin_id",
          "destination_id",
          "departure_at",
          "arrival_at",
          "price_cents",
          "total_capacity"
        ],
        "title": "FlightCreate"
      },
      "FlightPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "flight_number": {
            "type": "string",
            "title": "Flight Number"
          },
          "origin": {
            "$ref": "#/components/schemas/AirportPublic"
          },
          "destination": {
            "$ref": "#/components/schemas/AirportPublic"
          },
          "departure_at": {
            "type": "string",
            "format": "date-time",
            "title": "Departure At"
          },
          "arrival_at": {
            "type": "string",
            "format": "date-time",
            "title": "Arrival At"
          },
          "price_cents": {
            "type": "integer",
            "title": "Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "total_capacity": {
            "type": "integer",
            "title": "Total Capacity"
          },
          "available_capacity": {
            "type": "integer",
            "title": "Available Capacity"
          },
          "status": {
            "$ref": "#/components/schemas/FlightStatus"
          },
          "effective_status": {
            "$ref": "#/components/schemas/EffectiveFlightStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "flight_number",
          "origin",
          "destination",
          "departure_at",
          "arrival_at",
          "price_cents",
          "currency",
          "total_capacity",
          "available_capacity",
          "status",
          "effective_status"
        ],
        "title": "FlightPublic"
      },
      "FlightStatus": {
        "type": "string",
        "enum": ["scheduled", "cancelled"],
        "title": "FlightStatus"
      },
      "FlightUpdate": {
        "properties": {
          "flight_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 12,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Flight Number"
          },
          "origin_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Id"
          },
          "destination_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Id"
          },
          "departure_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure At"
          },
          "arrival_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival At"
          },
          "price_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Cents"
          },
          "total_capacity": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Capacity"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlightStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "FlightUpdate"
      },
      "FlightsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FlightPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "FlightsPublic"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "Message"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}